factory.py refactoring
